Extract base class from code actions tests
We were running most of them twice. Very confusing when you expect one test failure, but get two.
